requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
narwhals==2.0.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.97.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pipecat-ai @ git+https://github.com/pipecat-ai/pipecat.git@main
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rpds-py==0.26.0
scipy==1.16.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
torch==2.2.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==14.2
yarl==1.20.1
